Add a Check authentication button
The check authentication button is associated with a label that is bound
to a property of the main object.

We added a simple check_auth method to both the NdfApp and Connection
objects.
NdfApp.check_auth toggles its check_auth_token property
Connection.check_auth launches a query against the /login path but
doesn't stores the cookie after connecting.